Fix: move critical functions
